we move cursor in state
