+++ scripts/hello-node/game-reducer.js
@@ -1,25 +1,40 @@
-const R = require('ramda');
+const R = require('ramda')
+const GameActions = require('./actions')
+const BOARD_SIZE = 3
 
 const createEmptyBoard = () =>
-  R.range(0, 9);
+  R.range(0, BOARD_SIZE * BOARD_SIZE)
 
 const isEven = x => x % 2 === 0;
+
 const createBoardWithMarkers = () =>
-  R.range(0, 9).map(R.ifElse(isEven, R.always('a'), R.always('b')));
+  R.range(0, BOARD_SIZE * BOARD_SIZE)
+    .map(R.ifElse(isEven, R.always('a'), R.always('b')));
 
-
-const createInitialState = () => (
-  {
-    cursorPos: [0, 0],
-    // board: createEmptyBoard()
-    board: createBoardWithMarkers()
-  }
-);
+const createInitialState = () => ({
+  cursorPos: [0, 0],
+  // board: createEmptyBoard()
+  board: createBoardWithMarkers()
+});
 
 const initialState = createInitialState();
 
-const gameReducer = (prevState = initialState, action) => {
-  return prevState;
+let clampWithinBoard = R.clamp(0, BOARD_SIZE - 1)
+const moveWithinBoard = (xfn, yfn) =>
+  R.compose(
+    R.adjust(R.compose(clampWithinBoard, xfn), 0),
+    R.adjust(R.compose(clampWithinBoard, yfn), 1)
+  );
+
+const gameReducer = (state = initialState, action) => {
+  switch (action.type) {
+    case GameActions.MOVE_CURSOR:
+      return Object.assign({}, state, {
+        cursorPos: moveWithinBoard(action.xfn, action.yfn)(state.cursorPos)
+      });
+    default:
+      return state
+  }
 };
 
 module.exports = {
