+++ scripts/hello-node/game.js
@@ -1,6 +1,5 @@
 const R = require('ramda')
 
-const BOARD_SIZE = 3
 
 
 
@@ -12,12 +11,7 @@ const playerAtGrid = (state, {row, col}) =>
 
 
 
-let clampWithinBoard = R.clamp(0, BOARD_SIZE - 1)
-const moveWithinBoard = (xfn, yfn) =>
-  R.compose(
-    R.adjust(R.compose(clampWithinBoard, xfn), 0),
-    R.adjust(R.compose(clampWithinBoard, yfn), 1)
-  )
+
 
 module.exports = {
   BOARD_SIZE,
